Moved scan settings model classes as per review comment.
